Created the Pokemon, Player and ProfessorOak header and source files
Created the Pokemon, Player and ProfessorOak header files in the Header Files folder. Created the Pokemon, Player and ProfessorOak classes in the respective header files. Declared the necessary functions in respective classes.

Created the Pokemon, Player and ProfessorOak source files in the Source Files folder. Defined the functions declared in the header files, in the respective source files.

Deleted their class definitions from main.cpp.
